replace u{1,2,4,8} with ui{8,16,32,64} types

+++ bio4d/fs_bc160.c
@@ -850,7 +850,7 @@ fs_bc160_put_bytes(struct request *r)
 	if (r->scan_size > 0) {
 
 		//  Note: regress, sanity test ... remove later.
-		if ((u8)r->scan_size != r->blob_size)
+		if ((i64)r->scan_size != r->blob_size)
 			_panic(r, "r->scan_size != r->blob_size");
 
 		if (r->scan_size > (int)(sizeof chunk - 1)) {
