+++ bio4d/fs_btc20.c
@@ -886,7 +886,7 @@ fs_btc20_put_bytes(struct request *r)
 	if (r->scan_size > 0) {
 
 		//  Note: regress, sanity test ... remove later.
-		if ((u8)r->scan_size != r->blob_size)
+		if ((i64)r->scan_size != r->blob_size)
 			_panic(r, "r->scan_size != r->blob_size");
 
 		if (r->scan_size > (int)(sizeof chunk - 1)) {
